refactor: minor clean up
